+++ inst/doc/kmedoid.R
@@ -23,6 +23,17 @@ result <- fastkmed(mrwdist, ncluster = 3, iterate = 50)
 (fastiris <- table(result$cluster, iris[,5]))
 (misclass <- (1-sum(diag(fastiris))/length(iris[,5]))*100)
 
+## ---- fig.width=7, fig.asp=0.8-------------------------------------------
+rownames(mrwdist) <- colnames(mrwdist) <- 1:nrow(mrwdist)
+siliris <- silhoutte(mrwdist, result$medoid, result$cluster)
+siliris$result[c(1:3,70:75,101:103),]
+siliris$plot
+
+## ---- fig.width=7, fig.asp=0.8-------------------------------------------
+shairis <- shadow(mrwdist, result$medoid, result$cluster)
+shairis$result[c(1:3,70:75,101:103),]
+shairis$plot
+
 ## ------------------------------------------------------------------------
 k <- 3
 # a simple and fast k-medoids function for bootstrap evaluation
@@ -52,10 +63,24 @@ wardorder <- function(x, nclust) {
 consensusfastkmed <- consensusmatrix(fastkmedboot, nclust = k, wardorder)
 consensusfastkmed[c(1:5,51:55,101:105),c(1:5,51:55,101:105)]
 
-## ------------------------------------------------------------------------
+## ---- fig.width=7, fig.asp=0.8-------------------------------------------
 clustheatmap(consensusfastkmed, "Iris Data via Fast K-medoids")
 
-## ------------------------------------------------------------------------
+## ---- fig.width=7, fig.asp=0.8-------------------------------------------
 consensuskmeans <- consensusmatrix(kmeansboot, nclust = k, wardorder)
 clustheatmap(consensuskmeans, "Iris Data via K-means")
+
+## ---- fig.width=7, fig.asp=0.8-------------------------------------------
+pcadat <- prcomp(iris[,1:4], scale. = TRUE)
+pcabiplot(pcadat, colobj = result$cluster)
+
+## ---- fig.width=7, fig.asp=0.8-------------------------------------------
+pcadat <- prcomp(iris[,1:4], scale. = TRUE)
+pcabiplot(pcadat, y = "PC3",colobj = result$cluster)
+
+## ---- fig.width=7, fig.asp=0.8-------------------------------------------
+barplotnum(iris[,1:4], result$cluster)
+
+## ---- fig.width=7, fig.asp=0.8-------------------------------------------
+barplotnum(iris[,1:4], result$cluster, nc = 2)