+++ NEWS.md
@@ -1,3 +1,20 @@
+# kmed 0.2.0
+
+* Added barplotnum to make a barplot of numerical data set.
+* Added pcabiplot to make biplot of pca class.
+* Added silhoutte to obtain silhoutte index and plot.
+* Added shadow to obtain shadow value index from medoid instead of centroid and its plot.
+* Added two data sets.
+* Fixed cooccur.
+* Fixed clustboot.
+* Fixed fastkmed.
+* Fixed matching.
+* Fixed stepkmed.
+* Fixed rankkmed.
+* Deleted coocurance.
+* Edited `NEWS.md`.
+* Edited vignette.
+
 # kmed 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.